Changed price type to double. Added console clearing. Introduced logic for basket subtotal

+++ RecruitmentTaskPromise/Product.cs
@@ -9,9 +9,9 @@ namespace RecruitmentTaskPromise
     internal class Product
     {
         public string Name { get; set; }
-        public int Price { get; set; }
+        public double Price { get; set; }
 
-        public Product(string name, int price)
+        public Product(string name, double price)
         {
             Name = name;
             Price = price;
